requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
configparser==7.0.0
cryptography==43.0.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cacheops==7.0.2
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-heroku==0.3.1
django-phonenumber-field==8.0.0
django-rest-auth==0.9.5
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ez_setup==0.9
funcy==2.0
gunicorn==22.0.0
idna==3.7
mysql-connector-python==9.0.0
mysqlclient==2.1.1
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stripe==10.6.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
